separate server_init_db_queries() from termdb.sql which imports app.js and causes the updated tests to launch unnecessary server process
