+++ server/src/mds2.init.js
@@ -3,7 +3,7 @@ const path = require('path')
 const fs = require('fs').promises
 const spawn = require('child_process').spawn
 const utils = require('./utils')
-const server_init_db_queries = require('./termdb.sql').server_init_db_queries
+const server_init_db_queries = require('./termdb.server.init').server_init_db_queries
 const validate_single_numericrange = require('#shared/mds.termdb.termvaluesetting').validate_single_numericrange
 const serverconfig = require('./serverconfig')
 const { setDbRefreshRoute } = require('./dsUpdateAttr.js')
