+++ server/src/mds3.init.js
@@ -7,7 +7,8 @@ const utils = require('./utils')
 const compute_mclass = require('./vcf.mclass').compute_mclass
 const serverconfig = require('./serverconfig')
 const { dtsnvindel, dtfusionrna, dtsv, mclassfusionrna, mclasssv } = require('#shared/common')
-const { get_samples, server_init_db_queries } = require('./termdb.sql')
+const { get_samples } = require('./termdb.sql')
+const { server_init_db_queries } = require('./termdb.server.init')
 const { barchart_data } = require('./termdb.barchart')
 const { setDbRefreshRoute } = require('./dsUpdateAttr.js')
 const mayInitiateScatterplots = require('./termdb.scatter').mayInitiateScatterplots
